# qremergencias-ws/src/main/resources/swagger.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/qremergencias","tags":[{"name":"qr-emergencias-controller","description":"QR Emergencias Controller"},{"name":"user-front-controller","description":"User Front Controller"}],"paths":{"/api/token":{"get":{"tags":["qr-emergencias-controller"],"summary":"getToken","operationId":"getTokenUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"string"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/userFront/changePassword":{"post":{"tags":["user-front-controller"],"summary":"changePassword","operationId":"changePasswordUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"changePassword","description":"changePassword","required":false,"schema":{"$ref":"#/definitions/ChangePasswordDTO"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/userFront/register":{"post":{"tags":["user-front-controller"],"summary":"register","operationId":"registerUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"model","description":"model","required":false,"schema":{"$ref":"#/definitions/CreateUserDTO"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/LoginUserDTO"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/userFront/resetPassword":{"post":{"tags":["user-front-controller"],"summary":"resetPassword","operationId":"resetPasswordUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"resetPassword","description":"resetPassword","required":false,"schema":{"$ref":"#/definitions/ResetPasswordDTO"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/userFront/sendMailConfirmation":{"post":{"tags":["user-front-controller"],"summary":"sendMailConfirmation","operationId":"sendMailConfirmationUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"g-recaptcha-response","in":"query","description":"g-recaptcha-response","required":true,"type":"string"},{"name":"username","in":"query","description":"username","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"LoginUserDTO":{"type":"object","properties":{"imageUrl":{"type":"string"},"lastName":{"type":"string"},"name":{"type":"string"}}}}}